Apply suggestions from code review and reorganize test utils
